LLM_Project/requirements.txt
openai
langchain>=0.1.0
python-dotenv
tqdm